web/requirements.txt
Django==4.0.7
djangorestframework==3.13.1
django-filter==22.1
django-widget-tweaks==1.4.12
gnupg==2.3.1
Markdown==3.4.1
python-apt @ git+https://salsa.debian.org/apt-team/python-apt.git@6145b6484d220685edfd922d364afbf065127efe
python-dateutil==2.8.2
python-gnupg==0.5.0
pytz==2022.2.1
requests==2.28.1
rpmfile==1.0.8
sqlparse==0.4.2
urllib3==1.26.11
